fix: Исправил прошлые ошибки

+++ app/Http/Controllers/admin/DestroyUserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class DestroyUserController extends Controller
 {
@@ -13,7 +14,8 @@ class DestroyUserController extends Controller
      */
     public function __invoke(User $user)
     {
+        $this->authorize('delete', $user);
         $user->delete();
-        return redirect()->route('admin.index');
+        return Auth::user()->getAuthIdentifier() !== $user->getKey() ? redirect()->route('admin.index') : redirect()->route('register');
     }
 }
